dps: remove usage of manually listed fights
There was some thought, from ancient times, back when there were dps
checks like a3s hand of pain, that it'd be useful to have per-fight
phase listings.  This was also used prior to 0b47777f904f98a2 where
the pullcounter will auto-start in parties of 8.

The only thing that these do for the dps meter is to give it a title,
which is cute, but not worth the complication this code adds.